# :+1: Use Web Streams API instead of Deno.Reader/Deno.Writer to follow recent deno favor