--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,3 @@
-/.tools
 .tool-versions
+deno.lock
+.coverage