--- a/denops/@denops-private/impl_test.ts
+++ b/denops/@denops-private/impl_test.ts
@@ -3,7 +3,7 @@ import {
   assertEquals,
   assertRejects,
 } from "https://deno.land/std@0.186.0/testing/asserts.ts";
-import { test } from "https://deno.land/x/denops_test@v1.2.0/mod.ts#^";
+import { test } from "https://deno.land/x/denops_test@v1.3.1/mod.ts#^";
 import { BatchError } from "../@denops/mod.ts";
 
 test({