--- a/denops/@denops-private/impl_test.ts
+++ b/denops/@denops-private/impl_test.ts
@@ -1,8 +1,8 @@
-import * as path from "https://deno.land/std@0.186.0/path/mod.ts";
+import * as path from "https://deno.land/std@0.187.0/path/mod.ts";
 import {
   assertEquals,
   assertRejects,
-} from "https://deno.land/std@0.186.0/testing/asserts.ts";
+} from "https://deno.land/std@0.187.0/testing/asserts.ts";
 import { test } from "https://deno.land/x/denops_test@v1.3.1/mod.ts#^";
 import { BatchError } from "../@denops/mod.ts";