--- a/denops/@denops-private/service.ts
+++ b/denops/@denops-private/service.ts
@@ -1,4 +1,4 @@
-import { toFileUrl } from "https://deno.land/std@0.186.0/path/mod.ts";
+import { toFileUrl } from "https://deno.land/std@0.187.0/path/mod.ts";
 import {
   assertArray,
   assertBoolean,