--- a/denops/@denops-private/worker/script.ts
+++ b/denops/@denops-private/worker/script.ts
@@ -4,13 +4,14 @@ import {
   isObject,
   isString,
 } from "https://deno.land/x/unknownutil@v2.1.1/mod.ts#^";
-import { Session } from "https://deno.land/x/msgpack_rpc@v3.1.6/mod.ts#^";
-import { using } from "https://deno.land/x/disposable@v1.1.1/mod.ts#^";
 import {
-  WorkerReader,
-  WorkerWriter,
+  Client,
+  Session,
+} from "https://deno.land/x/messagepack_rpc@v0.3.1/mod.ts#^";
+import {
+  readableStreamFromWorker,
+  writableStreamFromWorker,
 } from "https://deno.land/x/workerio@v3.1.0/mod.ts#^";
-import { responseTimeout } from "../defs.ts";
 import type { Denops, Meta } from "../../@denops/mod.ts";
 import { DenopsImpl } from "../impl.ts";
 import { patchConsole } from "./patch_console.ts";
@@ -21,72 +22,89 @@ async function main(
   scriptUrl: string,
   meta: Meta,
 ): Promise<void> {
-  const reader = new WorkerReader(worker);
-  const writer = new WorkerWriter(worker);
-  await using(
-    new Session(reader, writer, {}, {
-      responseTimeout,
-      errorCallback(e) {
-        if (e.name === "Interrupted") {
-          return;
-        }
-        console.error("Unexpected error occurred", e);
-      },
-    }),
-    async (session) => {
-      // Protect the process itself from "Unhandled promises"
-      // https://github.com/vim-denops/denops.vim/issues/208
-      globalThis.addEventListener("unhandledrejection", (ev) => {
-        let { reason } = ev;
-        if (reason instanceof Error && reason.stack) {
-          reason = reason.stack;
-        }
-        console.error(
-          `Unhandled rejection is detected. Worker will be reloaded: ${reason}`,
-        );
-        // Reload the worker because "Unhandled promises" error occured.
-        session.notify("reload");
-        // Avoid process death
-        ev.preventDefault();
-      });
-      const denops: Denops = new DenopsImpl(worker.name, meta, session);
-      try {
-        // Import module with fragment so that reload works properly
-        // https://github.com/vim-denops/denops.vim/issues/227
-        const mod = await import(`${scriptUrl}#${performance.now()}`);
-        await denops.cmd(
-          `doautocmd <nomodeline> User DenopsSystemPluginPre:${worker.name}`,
-        )
-          .catch((e) =>
-            console.warn(
-              `Failed to emit DenopsSystemPluginPre:${worker.name}: ${e}`,
-            )
-          );
-        await mod.main(denops);
-        await denops.cmd(
-          `doautocmd <nomodeline> User DenopsSystemPluginPost:${worker.name}`,
-        )
-          .catch((e) =>
-            console.warn(
-              `Failed to emit DenopsSystemPluginPost:${worker.name}: ${e}`,
-            )
-          );
-      } catch (e) {
-        console.error(e);
-        await denops.cmd(
-          `doautocmd <nomodeline> User DenopsSystemPluginFail:${worker.name}`,
-        )
-          .catch((e) =>
-            console.warn(
-              `Failed to emit DenopsSystemPluginFail:${worker.name}: ${e}`,
-            )
-          );
-      } finally {
-        await session.waitClosed();
-      }
+  const session = new Session(
+    readableStreamFromWorker(worker),
+    writableStreamFromWorker(worker),
+    {
+      onRequestMessageError: onMessageError,
+      onResponseMessageError: onMessageError,
+      onNotificationMessageError: onMessageError,
     },
   );
+  session.start();
+  const client = new Client(session);
+  // Protect the process itself from "Unhandled promises"
+  // https://github.com/vim-denops/denops.vim/issues/208
+  globalThis.addEventListener("unhandledrejection", (ev) => {
+    let { reason } = ev;
+    if (reason instanceof Error && reason.stack) {
+      reason = reason.stack;
+    }
+    console.error(
+      `Unhandled rejection is detected. Worker will be reloaded: ${reason}`,
+    );
+    // Reload the worker because "Unhandled promises" error occured.
+    client.notify("reload");
+    // Avoid process death
+    ev.preventDefault();
+  });
+  const denops: Denops = new DenopsImpl(worker.name, meta, {
+    get dispatcher() {
+      return session.dispatcher;
+    },
+    set dispatcher(dispatcher) {
+      session.dispatcher = dispatcher;
+    },
+    call(method: string, ...params: unknown[]): Promise<unknown> {
+      return client.call(method, ...params);
+    },
+    notify(method: string, ...params: unknown[]): Promise<void> {
+      client.notify(method, ...params);
+      return Promise.resolve();
+    },
+  });
+  try {
+    // Import module with fragment so that reload works properly
+    // https://github.com/vim-denops/denops.vim/issues/227
+    const mod = await import(`${scriptUrl}#${performance.now()}`);
+    await denops.cmd(
+      `doautocmd <nomodeline> User DenopsSystemPluginPre:${worker.name}`,
+    )
+      .catch((e) =>
+        console.warn(
+          `Failed to emit DenopsSystemPluginPre:${worker.name}: ${e}`,
+        )
+      );
+    await mod.main(denops);
+    await denops.cmd(
+      `doautocmd <nomodeline> User DenopsSystemPluginPost:${worker.name}`,
+    )
+      .catch((e) =>
+        console.warn(
+          `Failed to emit DenopsSystemPluginPost:${worker.name}: ${e}`,
+        )
+      );
+    await session.wait();
+  } catch (e) {
+    console.error(e);
+    await denops.cmd(
+      `doautocmd <nomodeline> User DenopsSystemPluginFail:${worker.name}`,
+    )
+      .catch((e) =>
+        console.warn(
+          `Failed to emit DenopsSystemPluginFail:${worker.name}: ${e}`,
+        )
+      );
+    await session.shutdown();
+  }
   self.close();
+}
+
+function onMessageError(message: unknown, error: Error): void {
+  if (error instanceof Error && error.name === "Interrupted") {
+    return;
+  }
+  console.error(`Failed to handle message ${message}`, error);
 }
 
 function isMeta(v: unknown): v is Meta {