… denops/@denops-private/tee.ts
… denops/@denops-private/tee_test.ts